python/examples/extract_images_from_walk/requirements.txt
matplotlib>=3.7.2
numpy>=1.25.1
bosdyn-api>=3.3.0
python-pptx>=0.6.21